AQ2_makeBY_Dhruv0805/requirements.txt
flask==3.0.3
flask-cors==4.0.1
pandas==2.2.2
numpy==1.26.4
scikit-learn==1.5.1
joblib==1.4.2
openpyxl==3.1.5
python-dotenv==1.0.1
werkzeug==3.0.4
twilio==9.3.0
